backend/requirements_minimal.txt
Django==5.2.3
djangorestframework==3.16.0
drf-spectacular==0.29.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.0
PyJWT==2.8.0
pytest==8.3.3
pytest-django==4.9.0